+++ lib/collection/index.js
@@ -3,14 +3,15 @@ const path = require('path');
 const fs = require('fs-extra');
 Promise.promisifyAll(fs);
 const logger = require('winston');
-const config = require('./config');
-const CollectionPage = require('./collection-page');
-const CollectionPagination = require('./collection-pagination');
 const chunk = require('lodash/array/chunk');
 const each = require('lodash/collection/each');
 const reduce = require('lodash/collection/reduce');
 const sortBy = require('lodash/collection/sortBy');
 const isUndefined = require('lodash/lang/isUndefined');
+
+const config = require('../config');
+const CollectionPage = require('./page');
+const CollectionPagination = require('./pagination');
 
 class Collection {
   constructor(name = '', collectionConfig = {}) {
