+++ lib/collection/page.js
@@ -1,6 +1,6 @@
-const config = require('./config');
-const utils = require('./utils');
 const path = require('path');
+const config = require('../config');
+const utils = require('../utils');
 
 class CollectionPage {
   constructor(files, permalink, data) {
