+++ lib/collection/pagination.js
